Promote static methods to functions per Kotlin best practices
