+++ src/main/kotlin/tech/relaycorp/relaynet/wrappers/PRNG.kt
@@ -1,0 +1,8 @@
+package tech.relaycorp.relaynet.wrappers
+
+import java.security.SecureRandom
+
+fun generateRandom64BitValue(): Long {
+    val random = SecureRandom()
+    return random.nextLong()
+}
